requirements.txt
pandas
numpy
openpyxl
pymongo
dnspython
python-dotenv
